Fix missing curl binary and update CI
Fix parquet file download error because of a missing curl binary.
This was happening on the `eth_getBalance` test for instance.
Note that flood swallowed the exception making the error harder to spot.

Update the CI to also test against eth_getBalance which triggers a
parquet file download.
Also removes the redundant node2, this was a workaround as flood used to
crash with only one node.

Also unpin the patch version of Python and address a small DRY on the
Vegeta version.